backend/requirements-test.txt
# Base dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
python-dotenv==1.0.0
pycoingecko==3.1.0

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 